% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsem_formula.R
\name{ar}
\alias{ar}
\title{Title}
\usage{
ar(y, p = 1, data = data)
}
\arguments{
\item{y}{The variable in \code{data} used for the autoregressive process
(should be the same as the response variable).}

\item{p}{A non-negative integer specifying the autoregressive (AR) order.
Default is 1.}
}
\description{
Title
}
