% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1_data.R
\docType{data}
\name{ar1_data}
\alias{ar1_data}
\title{Simple Time-Series Data}
\format{
\subsection{\code{ar1_data}}{

A data frame with 2,500 rows and 3 columns:
\describe{
\item{ID}{Unit identifier}
\item{time}{Time point}
\item{Y1}{The time-series variable}
}
}
}
\source{
\code{\link{mlts_sim}}
}
\usage{
ar1_data
}
\description{
Simulated Data (from \code{\link{mlts_sim}}) for one time-series variable.
}
\keyword{datasets}
