% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARprepare.R
\name{ARprepare}
\alias{ARprepare}
\title{Title}
\usage{
ARprepare(
  VARmodel,
  data,
  ts.ind,
  covariates = NULL,
  outcomes = NULL,
  center.covs = T,
  std.outcome = T,
  outcome.pred.btw = NULL
)
}
\arguments{
\item{VARmodel}{data.frame. Output of VARmodel-Functions.}

\item{data}{data.frame. Data input.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
