% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_dsem_data}
\alias{create_dsem_data}
\title{Create data for DSEM models}
\usage{
create_dsem_data(data, delta, id, time)
}
\arguments{
\item{data}{An object of class \code{data.frame} (or one that can be coerced
to that class) containing data of all variables used in the model.}

\item{delta}{The step interval for approximation for a continuous time DSEM.
The smaller the step interval, the better the approximation.}

\item{id}{The variable in \code{data} that identifies the person or observational
unit (as string).}

\item{time}{The variable in \code{data} that contains the (continuous) time (as
string).}
}
\value{
A \code{data.frame} with missings imputed for use in \code{dsem()}.
}
\description{
Create data for DSEM models
}
\examples{
Tbd
}
