% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dummies.R
\name{create_dummies}
\alias{create_dummies}
\title{Create Dummy Variables}
\usage{
create_dummies(data, vars)
}
\arguments{
\item{data}{A data frame.}

\item{vars}{A character vector of column names. These columns will be transformed to 
a set of dummy variables. Both the original column and the first dummy variable are dropped.}
}
\value{
A data frame.
}
\description{
Creates a set of dummy variables from specified columns in a data frame.
}
\examples{
# transform the cyl column to a set of dummy variables
mtcars_cyl_dummies <- create_dummies(mtcars, "cyl")
head(mtcars_cyl_dummies)

}
