% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lags.R
\name{create_lags}
\alias{create_lags}
\title{Create Lag Variables}
\usage{
create_lags(data, lags, vars)
}
\arguments{
\item{data}{A data frame.}

\item{lags}{A numeric vector of lags.}

\item{vars}{A character vector of column names.}
}
\value{
A data frame.
}
\description{
Takes a data frame and adds new columns that are lags of existing columns.
}
\details{
For each column name, the specified lags are added as new columns to the data frame.
}
\examples{
# create lags of 2 and 4 of mpg
mtcars_mpg_lags <- create_lags(mtcars, c(2, 4), "mpg")
head(mtcars_mpg_lags)

}
