% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_missings.R
\name{create_missings}
\alias{create_missings}
\title{Create missings for DSEM data}
\usage{
create_missings(data, delta, id, time, btw.vars, clean = T)
}
\arguments{
\item{data}{An object of class \code{data.frame} (or one that can be coerced
to that class) containing data of all variables used in the model.}

\item{delta}{The step interval for approximation for a continuous time DSEM.
The smaller the step interval, the better the approximation.}

\item{id}{The variable in \code{data} that identifies the person or observational
unit (as character).}

\item{time}{The variable in \code{data} that contains the (continuous) time (as
string).}

\item{btw.vars}{The names of between-level variables in the data to be
added in newly created rows with NAs.}

\item{clean}{Remove helper columns .}
}
\value{
A \code{data.frame} with missings imputed for use in \code{dsem()}.
}
\description{
Create missings for DSEM data
}
\examples{
Tbd
}
