% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_missings.R
\name{create_missings3}
\alias{create_missings3}
\title{Create Missings for Approximation of Continuous Time Dynamic Models
(Deprecated)}
\usage{
create_missings3(data, tinterval, id, time, btw_vars = NULL, clean = TRUE)
}
\arguments{
\item{data}{An object of class \code{data.frame} (or one that can be coerced
to that class) containing data of all variables used in the model.}

\item{tinterval}{The step interval for approximation for a continuous time DSEM.
The smaller the step interval, the better the approximation.}

\item{id}{The variable in \code{data} that identifies the person or observational
unit (as character).}

\item{time}{The variable in \code{data} that contains the (continuous) time (as
string).}

\item{btw_vars}{The names of between-level variables in the data to be
added in newly created rows with NAs.}

\item{clean}{Logical. Remove helper columns. Defaults to \code{TRUE}.}
}
\value{
A \code{data.frame} with missings imputed for use in \code{prepare_data()}.
}
\description{
Create Missings for Approximation of Continuous Time Dynamic Models
(Deprecated)
}
