% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stan_data.R
\name{create_stan_data}
\alias{create_stan_data}
\title{Create data for Stan}
\usage{
create_stan_data(
  data,
  id,
  beep,
  ts,
  pred_random = NULL,
  outcome = NULL,
  out_predictors = NULL,
  standardize_out = TRUE,
  overnight_lags = NULL,
  miss_handling = c("remove", "impute"),
  random.innovations = TRUE,
  add_mplus_data = T
)
}
\arguments{
\item{data}{An object of class \code{data.frame} (or one that can be coerced
to that class) in long format.}

\item{id}{character. The variable in \code{data} that identifies the person or
observational unit.}

\item{beep}{character. The variable in \code{data} that contains the running
beep from 1 to TP for each person.}

\item{ts}{character. The variable in \code{data} that contains the observations
of the time-varying construct.}

\item{pred_random}{character. Name(s) of between-person variables to use as
predictors of individual parameters}

\item{outcome}{character. Names(s) of between-person variables
to regress individual parameters on.}

\item{out_predictors}{character. "mu", "ar", and/or "logv" to use individual
parameter estimates.
Can be combined with any other between-person variable.}

\item{standardize_out}{logical. Should output be standardized? Defaults
to \code{TRUE}.}

\item{overnight_lags}{character. Add column name of the day number to avoid
using the last observation of a day as lagged predictor of the
subsequent observation on the next day.}

\item{miss_handling}{character. Should missings be removed (\code{"remove"}) or
imputed (\code{"impute"})?}

\item{random.innovations}{logical. \code{FALSE} = constant innovation variance,
\code{TRUE} = person-specific innovation variances.}

\item{add_mplus_data}{logical. Should data be stored in Mplus format?
Defaults to \code{TRUE}.}
}
\value{
A \code{list} that can be passed to \code{stan()}.
}
\description{
Create data for Stan
}
\examples{
TBA
}
