% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhr_cv.R
\name{dhr_cv}
\alias{dhr_cv}
\title{Dynamic Harmonic Regression for Time Series with Cross-Validation}
\usage{
dhr_cv(y, h, lags = NULL, x = NULL, x_include = names(x),
  x_include_lags = NULL, x_include_dummy = NULL, order = c(0, 0, 0),
  s_order = c(0, 0, 0), s_period = 1, fourier_k = NULL,
  auto_season_dummy = F, drift = T, constant = T, mean = T,
  roll_update = F, method = "ML", seed = 1, verbose = F,
  out_tbl_width = 140)
}
\arguments{
\item{y}{A time series object (ts or msts).}

\item{h}{The number of periods for forecasting. If multiple values are provided, a rolling forecast origin is utilized.}

\item{lags}{Lags of the series. If provided, these will be added as regressors in the model.}

\item{x}{A data frame of external regressors. The number of rows must match the length of y.}

\item{x_include}{A vector of column names of x to include as regressors in the model. 
By default, all columns of x are included as regressors in the model.}

\item{x_include_lags}{A vector of column names of x where lags of these columns are added as regressors in the model. 
By default, only lags of the series are added as regressors when the lag argument is utilized. 
This provides the option to also include lags of the external regressors.}

\item{x_include_dummy}{A vector of column names of x where these columns are transformed to a set of dummy regressors.}

\item{order}{See order in \link[forecast]{Arima}}

\item{s_order}{See seasonal in \link[forecast]{Arima}}

\item{s_period}{The seasonal frequency.}

\item{fourier_k}{The number of fourier terms to add as regressors. 
This vector must have the same length as the number of seasonal periods of y, 
and each value must not exceed half of the size of the seasonal period.}

\item{auto_season_dummy}{If true, dummy variables for the seasonal periods are added as regressors.}

\item{drift}{See include.drift in \link[forecast]{Arima}}

\item{constant}{See include.constant in \link[forecast]{Arima}}

\item{mean}{See include.mean in \link[forecast]{Arima}}

\item{roll_update}{Option to update the model when the forecast origin is rolled forward.}

\item{seed}{Option to specify the seed.}

\item{verbose}{Option to print info to console.}

\item{out_tbl_width}{Option to set the info table width when verbose = T.}
}
\value{
A list. See the following details: \cr
 y - the original time series. \cr
 x_train - the subset of regressors used in training. \cr
 y_train - the subset of y used in training. \cr
 x_test - the subset of regressors used in testing. \cr
 y_test - the subset of y used in testing. \cr
 y_fitted - the fitted values (from the training stage). \cr
 y_hat - the predicted values (from the testing stage). \cr
 train_rmse - the training root mean squared error. \cr
 train_mape - the training mean absolute percentage error. \cr
 test_rmse - the testing (cross-validated) root mean squared error. \cr
 test_mape - the testing (cross-validated) mean absolute percentage error. \cr
 mod - the final model (an Arima object). \cr
 rt - the total runtime. \cr
}
\description{
This function fits a dynamic harmonic regression model to time series data
 and evaluates predictive performance using cross-validation.
}
