% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsem_ar.R
\name{dsem_ar}
\alias{dsem_ar}
\title{Title}
\usage{
dsem_ar(y, id, beep, data, miss_handling = "remove", iter = 2000, seed = NULL)
}
\arguments{
\item{y}{character. The variable in \code{data} that contains the observations
of the time-varying construct.}

\item{id}{character. The variable in \code{data} that identifies the person or
observational unit.}

\item{beep}{character. The variable in \code{data} that contains the running
beep from 1 to TP for each person.}

\item{data}{An object of class \code{data.frame} (or one that can be coerced
to that class) in long format.}

\item{miss_handling}{character. Should missings be removed (\code{"remove"}) or
imputed (\code{"impute"})?}

\item{iter}{Number of total iterations per chain (including warmup; defaults
to 2000).}

\item{seed}{The seed for random number generation to make results
reproducible. If \code{NULL} (the default), \pkg{Stan} will set the seed
randomly.}
}
\value{
An object of class \code{stanfit}.
}
\description{
Title
}
