% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsemr-package.R
\docType{package}
\name{dsemr-package}
\alias{dsemr}
\alias{dsemr-package}
\title{dsemr: Dynamic Structural Equation Models with R and Stan}
\description{
A package for fitting popular Dynamic Structural Equation Models (DSEM) with R and Stan.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/munchfab/dsemr}
  \item Report bugs at \url{https://github.com/munchfab/dsemr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
