% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplot.R
\name{fplot}
\alias{fplot}
\title{Forecast Plot}
\usage{
fplot(mlts_mod, start = 1, shade = T, colors = c("#347B98",
  "#FB9902"), x_label = "Time", y_label = "Series",
  color_labels = c("Observed", "Predicted"), title = "Forecast Plot",
  lsize = NULL, interactive = T)
}
\arguments{
\item{mlts_mod}{An output list from \link[MLTS]{dhr_cv} or \link[MLTS]{mlts_cv}.}

\item{start}{Starting time for the plot. By default, the entire series is plotted (start = 1).}

\item{shade}{Option to shade the forecast portion of the plot.}

\item{colors}{Option to set the line colors.}

\item{x_label}{Option to set the x-axis label.}

\item{y_label}{Option to set the y-axis label.}

\item{color_labels}{Option to set the color labels.}

\item{title}{Option to set the title.}

\item{lsize}{Option to set the line size.}

\item{interactive}{Option for an interactive plot.}
}
\value{
a ggplotly or plotly
}
\description{
Plot forecasts from \link[MLTS]{dhr_cv} or \link[MLTS]{mlts_cv} output.
}
