% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_plot.R
\name{importance_plot}
\alias{importance_plot}
\title{Importance Plot for XGBoost}
\usage{
importance_plot(xgb_importance, top_n = 25)
}
\arguments{
\item{xgb_importance}{Variable importance from an xgboost model. See \link[xgboost]{xgb.importance}.}

\item{top_n}{Number of variables to plot. Default is top_n = 25.}
}
\value{
A ggplot.
}
\description{
Plot variable importance from an xgboost model.
}
