% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts-package.R
\docType{package}
\name{mlts-package}
\alias{mlts}
\alias{mlts-package}
\title{mlts: Multilevel Latent Time Series Models with R and Stan}
\description{
Fit multilevel manifest or latent time-series models with R and Stan, including popular Dynamic Structural Equation Models (DSEM).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/munchfab/mlts}
  \item Report bugs at \url{https://github.com/munchfab/mlts/issues}
}

}
\author{
\strong{Maintainer}: Kenneth Koslowski \email{kenneth.koslowski@uni-leipzig.de}

Authors:
\itemize{
  \item Fabian Münch \email{fabian.felix.muench@uni-jena.de}
  \item Tobias Koch \email{tobias.koch@uni-jena.de}
  \item Jana Holtmann \email{jana.holtmann@uni-leipzig.de}
}

}
\keyword{internal}
