% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_cv.R
\name{mlts_cv}
\alias{mlts_cv}
\title{Machine Learning for Time Series with Cross-Validation}
\usage{
mlts_cv(y, method = "xgb", h, lags = max(h), x = NULL,
  x_include = names(x), x_include_lags = NULL,
  x_include_dummy = NULL, fourier_k = NULL, auto_season = F,
  params = NULL, roll_update = F, seed = 1, verbose = F,
  out_tbl_width = 140)
}
\arguments{
\item{y}{A time series object (ts or msts).}

\item{method}{The machine learning method. Options are one of ("xgb", "rf", "svm").}

\item{h}{The number of periods for forecasting. If multiple values are provided, a rolling forecast origin is utilized.}

\item{lags}{lags of the response to be used as predictors.}

\item{x}{A data frame of external features. The number of rows must match the length of y.}

\item{x_include}{A vector of column names of x to include as features in the model. 
By default, all columns of x are included as features in the model.}

\item{x_include_lags}{A vector of column names of x where lags of these columns are added as features in the model. 
By default, only lags of the series are added as features when the lag argument is utilized. 
This provides the option to also include lags of the external features.}

\item{x_include_dummy}{A vector of column names of x where these columns are transformed to a set of dummy features.}

\item{fourier_k}{The number of fourier terms to add as features. 
This vector must have the same length as the number of seasonal periods of y, 
and each value must not exceed half of the size of the seasonal period.}

\item{params}{Optional list of hyperparameter values for the selected machine learning method.}

\item{roll_update}{Option to update the model when the forecast origin is rolled forward.}

\item{seed}{Option to specify the seed.}

\item{verbose}{Option to print info to console.}

\item{out_tbl_width}{Option to set the info table width when verbose = T.}

\item{auto_season_dummy}{If true, dummy variables for the seasonal periods are added as features.}
}
\value{
A list. See the following details: \cr
 y - the original time series. \cr
 x_train - the subset of regressors used in training. \cr
 y_train - the subset of y used in training. \cr
 x_test - the subset of regressors used in testing. \cr
 y_test - the subset of y used in testing. \cr
 y_fitted - the fitted values (from the training stage). \cr
 y_hat - the predicted values (from the testing stage). \cr
 train_rmse - the training root mean squared error. \cr
 train_mape - the training mean absolute percentage error. \cr
 test_rmse - the testing (cross-validated) root mean squared error. \cr
 test_mape - the testing (cross-validated) mean absolute percentage error. \cr
 mod - the final model (an Arima object). \cr
 importance - variable importance if method = "xgb" or method - "rf". \cr
 rt - the total runtime. \cr
 params - the hyperparameters. \cr
}
\description{
This function fits a machine learning model to time series data
 and evaluates predictive performance using cross-validation.
}
