% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_model_betw.R
\name{mlts_model_betw}
\alias{mlts_model_betw}
\title{Title}
\usage{
mlts_model_betw(
  VARmodel,
  RE.pred = NULL,
  out.pred = NULL,
  out.pred.add.btw = NULL
)
}
\arguments{
\item{VARmodel}{data.frame. Output of \code{VARmodel()}-functions.}

\item{RE.pred}{character vector or a named list. Include between-level covariate(s)
as predictor(s) of all random effects in \code{VARmodel} by entering a vector of unique variable
names. Alternatively, to include between-level covariates or differing sets of
between-level covariates as predictors of specific random effects, a named
list (using the \code{param}-labels in \code{VARmodel}) can be entered (see details).
Note that if a named list is provided, all names that do not match random
parameters in \code{VARmodel} will be ignored.}

\item{out.pred}{character vector or a named list. Include between-level covariate(s)
as predictor(s) of all random effects in \code{VARmodel} by entering a vector of unique variable
names. Alternatively, to include between-level covariates or differing sets of
between-level covariates as predictors of specific random effects, a named
list (using the \code{param}-labels in \code{VARmodel}) can be entered (see details).}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
