% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_model_check.R
\name{mlts_model_check}
\alias{mlts_model_check}
\title{Helper function to run a series of user input checks and print detailed warnings}
\usage{
mlts_model_check(model, data, ts, covariates, outcomes)
}
\arguments{
\item{model}{data.frame. Output of model-Functions.}

\item{data}{data.frame. Data input.}

\item{ts}{tba}

\item{covariates}{tba}

\item{outcomes}{tba.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Helper function to run a series of user input checks and print detailed warnings
}
