% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_model_eval.R
\name{mlts_model_eval}
\alias{mlts_model_eval}
\title{Evaluate User Input from mlts model}
\usage{
mlts_model_eval(model)
}
\arguments{
\item{model}{data.frame. Output of \code{\link[mlts]{mlts_model}}.}
}
\value{
A \code{list}-object with details of the mlts model object.
}
\description{
Evaluate User Input from mlts model
}
\details{
Primarily for internal use.
}
\examples{
# fit simple autoregressive mlts model
ar_model <- mlts_model(q = 1)

# evaluate model infos
infos <- mlts_model_eval(ar_model)

}
