% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_model_formula.R
\name{mlts_model_formula}
\alias{mlts_model_formula}
\title{Create TeX Model Formula from VARmodel object}
\usage{
mlts_model_formula(VARmodel, data = NULL, labels = NULL)
}
\arguments{
\item{VARmodel}{The VARmodel object.}

\item{data}{An optional \code{data.frame} including the variables to be used
in the formula.}

\item{labels}{An optional character string including the names to be used
in the formula.}
}
\value{
An Rmarkdown file that is automatically rendered to a pdf document.
}
\description{
Create TeX Model Formula from VARmodel object
}
\examples{
1+1
}
