% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_model_priors.R
\name{mlts_model_priors}
\alias{mlts_model_priors}
\title{Add custom prior distributions to mlts model}
\usage{
mlts_model_priors(model, default = F)
}
\arguments{
\item{model}{data.frame. Output of \code{\link[mlts]{mlts_model}}.}

\item{default}{logical. If set to \code{TRUE}, default prior specifications are
added.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Add custom prior distributions to mlts model
}
