% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_plot.R
\name{mlts_plot}
\alias{mlts_plot}
\title{Plot results of mlts}
\usage{
mlts_plot(
  fit,
  type = c("fe", "re", "re.cor"),
  bpe = c("median", "mean"),
  sort_est = NULL,
  xlab = NULL,
  ylab = NULL,
  facet_ncol = 1,
  dot_size = 1,
  dot_color = "black",
  dot_shape = 1,
  errorbar_color = "black",
  errorbar_width = 0.3,
  add_true = FALSE,
  true_color = "red",
  true_shape = 22,
  true_size = 1,
  hide_xaxis_text = T,
  par_labels = NULL,
  labels_as_expressions = F
)
}
\arguments{
\item{fit}{An object of class \code{mlts.fit}}

\item{type}{Type of plot.
type = "fixef" (Default)
Forest-plot of model coefficients.
type = "ranef"
Plot of individual (random) effects
type = "raneftab"
(not yet implemented)}

\item{bpe}{The Bayesian point estimate is, by default, the median of the
posterior distribution (\code{bpe = "median"}). Set \code{bpe = "mean"} to use
the mean of the posterior distribution as point estimates.}

\item{sort_est}{Add parameter label for sorting of random effects.}

\item{xlab}{Title for the x axis.}

\item{ylab}{Title for the y axis.}

\item{facet_ncol}{Number of facet columns (see \code{ggplot2::facet_grid}).}

\item{dot_size}{Numeric, size of the dots that indicate the point estimates.}

\item{dot_color}{Character vector, indicating the color of the point estimates.}

\item{dot_shape}{Numeric, shape of the dots that indicate the point estimates.}

\item{errorbar_color}{integer.}

\item{errorbar_width}{integer.}

\item{add_true}{integer.}

\item{true_color}{integer.}

\item{true_shape}{integer.}

\item{true_size}{integer.}

\item{hide_xaxis_text}{logical.}

\item{par_labels}{integer.}

\item{labels_as_expressions}{logical.}
}
\value{
Returns a \code{ggplot}-object .
}
\description{
Plot results of mlts
}
\examples{
1 + 1
}
