% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_standardized.R
\name{mlts_standardized}
\alias{mlts_standardized}
\title{Get Standardized Estimates for an mlts Model}
\usage{
mlts_standardized(object, digit = 3, prob = 0.95)
}
\arguments{
\item{object}{\code{mltsfit}. Output of \code{\link[mlts]{mlts_model}} and related functions.}

\item{prob}{A value between 0 and 1 to indicate the width of the credible
interval. Default is .95.}

\item{digits}{Number of digits.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Get Standardized Estimates for an mlts Model
}
\examples{
\donttest{
# build simple vector-autoregressive mlts model for two time-series variables
var_model <- mlts_model(q = 2)

# fit model with (artificial) dataset ts_data
fit <- mlts_fit(
  model = var_model,
  data = ts_data,
  ts = c("Y1", "Y2"), # time-series variables
  id = "ID", # identifier variable
  tinterval = 1 # interval for approximation of continuous-time dynamic model,
)

# inspect standardized parameter estimates
mlts_standardized(fit)
}

}
