% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim.R
\name{plot_sim}
\alias{plot_sim}
\title{Plot Simulation Forecasts}
\usage{
plot_sim(sim, interactive = T)
}
\arguments{
\item{sim}{Output list from a simulation produced by \link[MLTS]{sim}.}

\item{interactive}{Option to make the plot interactive.}
}
\value{
A list of ggplot or ggplotly objects.
}
\description{
Plot Simulation Forecasts
}
