% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_runtime.R
\name{print_runtime}
\alias{print_runtime}
\title{Print Simulation Runtimes}
\usage{
print_runtime(out)
}
\arguments{
\item{out}{An output list from a simulation produced by \link[MLTS]{sim}.}
}
\value{
A tibble of runtimes.
}
\description{
Print Simulation Runtimes
}
