% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_sim.R
\name{print_sim}
\alias{print_sim}
\title{Print Forecast Simulation Results}
\usage{
print_sim(sim, scipen = 999, digits = 4,
  caption = "Simulation Results")
}
\arguments{
\item{sim}{An output list from a simulation produced by \link[MLTS]{sim}.}

\item{scipen}{Option to set scipen.}

\item{digits}{Option to set number of digits for rounding.}

\item{caption}{Option to add table caption.}
}
\value{
A tibble of simulation results.
}
\description{
Print Forecast Simulation Results
}
