% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_sims.R
\name{print_sims}
\alias{print_sims}
\title{Print Results from Multiple Simulations}
\usage{
print_sims(out)
}
\arguments{
\item{out}{An output list from multiple simulations produced by \link[MLTS]{run_sims}.}
}
\value{
A tibble of simulation results.
}
\description{
Print Results from Multiple Simulations
}
