% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sims.R
\name{run_sims}
\alias{run_sims}
\title{Run Multiple Forecast Simulations}
\usage{
run_sims(n_obs, n_rep, n_test, periods, arima_period, orders, ar, ma,
  k = NULL, lags = NULL, xgb_params, rf_params, svm_params, ml_k = k,
  auto_season = F, xreg_corr = NULL, burn = 100, method = "ML",
  seed = 1234)
}
\arguments{
\item{n_obs}{Length of the simulated time series.}

\item{n_rep}{Number of replicate time series.}

\item{n_test}{Length of the simulated time series held out as a test set.}

\item{periods}{Seasonal periods for the simualted time series.}

\item{arima_period}{Seasonal period used when fitting the ARIMA model.}

\item{orders}{See orders in \link[smooth]{ssarima}.}

\item{ar}{See AR in \link[smooth]{ssarima}.}

\item{ma}{See MA in \link[smooth]{ssarima}.}

\item{k}{Number of fourier terms included as predictors.}

\item{lags}{Lags of the simulated time series included as regressors in the ARIMA and DHR models.}

\item{xgb_params}{List of hyperparameters for the xgboost model.}

\item{rf_params}{List of hyperparameters for the randomForest model.}

\item{svm_params}{List of hyperparameters for the svm model.}

\item{ml_k}{Number of fourier terms included as features in the machine learning models.}

\item{xreg_corr}{Option to add covariates with a specified correlation with the series.}

\item{burn}{Option to use a burn-in when simulated the time series data.}

\item{seed}{Option to set the seed.}
}
\value{
List of root mean squared errors and runtimes for each method.
}
\description{
Generates multiple simulations by making repeated calls to \link[MLTS]{sim}.
}
