% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_data.R
\docType{data}
\name{ts_data}
\alias{ts_data}
\title{Simple Time-Series Data}
\format{
\subsection{\code{ts_data}}{

A data frame with 1,100 rows and 4 columns:
\describe{
\item{ID}{Unit identifier}
\item{time}{Time point}
\item{Y1, Y2}{The two time-series variables}
}
}
}
\source{
\code{\link{mlts_sim}}
}
\usage{
ts_data
}
\description{
Simulated Time-Series Data (from \code{\link{mlts_sim}}) for two
time-series variables.
}
\keyword{datasets}
