% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARfit.R
\name{VARfit}
\alias{VARfit}
\title{Title}
\usage{
VARfit(
  VARmodel,
  data,
  printMessage = T,
  printWarning = T,
  ts.ind,
  covariates = NULL,
  outcomes = NULL,
  center.covs = T,
  std.outcome = T,
  iter = 500,
  chains = 2,
  cores = 2,
  monitor.person.pars = F,
  fit.model = T,
  ...
)
}
\arguments{
\item{VARmodel}{data.frame. Output of VARmodel-Functions.}

\item{data}{data.frame. Data input.}

\item{printMessage}{logical. Print messages based on defined inputs (default = TRUE).}

\item{fit.model}{logical. Set to FALSE to avoid fitting the model which may be
helpful to inspect prepared data used for model estimation (default = T).}

\item{...}{Additional arguments passed to rstan \code{sampling}-function.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
