% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARmodelBuild.R
\name{VARmodelBuild}
\alias{VARmodelBuild}
\title{Title}
\usage{
VARmodelBuild(q, p = NULL, maxLag = c(1, 2, 3))
}
\arguments{
\item{q}{integer. The number of time-varying constructs.}

\item{p}{integer. For multiple-indicator models, specify a vector of length
\code{q} with the number of manifest indicators per construct.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
