% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARmodelChecks.R
\name{VARmodelchecks}
\alias{VARmodelchecks}
\title{Title}
\usage{
VARmodelchecks(VARmodel, data, ts.ind, covariates)
}
\arguments{
\item{VARmodel}{data.frame. Output of VARmodel-Functions.}

\item{data}{data.frame. Data input.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
