% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARmodelConstraints.R
\name{VARmodelConstraints}
\alias{VARmodelConstraints}
\title{Title}
\usage{
VARmodelConstraints(
  VARmodel,
  FixDynamics = F,
  FixInnoVars = F,
  FixInnoCovs = F,
  InnoCovsZero = F,
  FEis0 = NULL,
  REis0 = NULL
)
}
\arguments{
\item{VARmodel}{data.frame. Output of VARmodel-Functions.}

\item{FEis0}{character. A character vector to indeVARmodel which fiVARmodeled model parameters
should be fiVARmodeled to zero (Note: this results in removing the random effect
variance of the respective parameter).}

\item{REis0}{logical. Set to TRUE to treat all innovations as independent.}

\item{FiVARmodeledInnoVars}{logical. FiVARmodel all random effect variances (eVARmodelcept those
of individual traits) to zero.}

\item{FiVARmodeledCovs}{logical. Set all innovation covariances to a constant value.}

\item{CovsZero}{logical. Set to TRUE to treat all innovations as independent.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
