% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARmodelEval.R
\name{VARmodelEval}
\alias{VARmodelEval}
\title{Title}
\usage{
VARmodelEval(VARmodel)
}
\arguments{
\item{VARmodel}{data.frame. Output of VARmodel-Functions.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
