% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARmodelformula.R
\name{VARmodelformula}
\alias{VARmodelformula}
\title{Create TeX Model Formula from VARmodel object}
\usage{
VARmodelformula(VARmodel, data = NULL, labels = NULL)
}
\arguments{
\item{VARmodel}{The VARmodel object.}

\item{data}{An optional \code{data.frame} including the variables to be used in the formula.}

\item{labels}{An optional character string inluding the names to be used in the path diagram.}
}
\value{
An Rmarkdown file that is automatically rendered to a pdf document.
}
\description{
Create TeX Model Formula from VARmodel object
}
