% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARmodelMeasurement.R
\name{VARmodelMeasurement}
\alias{VARmodelMeasurement}
\title{Title}
\usage{
VARmodelMeasurement(VARmodel, q, p)
}
\arguments{
\item{VARmodel}{data.frame. Output of VARmodel-Functions.}

\item{q}{integer. The number of time-varying constructs.}

\item{p}{integer. A vector of length \code{q} with the number of manifest
indicators per construct.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
