% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARmodelPriors.R
\name{VARmodelPriors}
\alias{VARmodelPriors}
\title{Title}
\usage{
VARmodelPriors(VARmodel, default = F)
}
\arguments{
\item{VARmodel}{data.frame. Output of VARmodel-Functions.}

\item{default}{logical. If set to \code{TRUE}, default prior specifications are
added.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
