% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARmodelSim.R
\name{VARmodelSim}
\alias{VARmodelSim}
\title{Title}
\usage{
VARmodelSim(
  VARmodel,
  default = F,
  N,
  TP,
  burn.in = 500,
  seed = NULL,
  btw.var.sds = NULL
)
}
\arguments{
\item{VARmodel}{data.frame. Output of VARmodel-Functions.}

\item{default}{logical. If set to \code{TRUE}, default prior specifications are
added.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
