% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts-package.R
\docType{package}
\name{mlts-package}
\alias{mlts}
\alias{mlts-package}
\title{mlts: Dynamic Structural Equation Models with R and Stan}
\description{
A package for fitting popular Dynamic Structural Equation Models (DSEM) with R and Stan.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/munchfab/mlts}
  \item Report bugs at \url{https://github.com/munchfab/mlts/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
