% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts-package.R
\docType{package}
\name{mlts-package}
\alias{mlts}
\alias{mlts-package}
\title{mlts: Multilevel Latent Time Series Models with 'R' and 'Stan'}
\description{
Fit multilevel manifest or latent time-series models, including popular Dynamic Structural Equation Models (DSEM). The models can be set up and modified with user-friendly functions and are fit to the data using 'Stan' for Bayesian inference. Path models and formulas for user-defined models can be easily created with functions using 'knitr'. Asparouhov, Hamaker, & Muthen (2018) \doi{10.1080/10705511.2017.1406803}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/munchfab/mlts}
  \item Report bugs at \url{https://github.com/munchfab/mlts/issues}
}

}
\author{
\strong{Maintainer}: Kenneth Koslowski \email{kenneth.koslowski@uni-leipzig.de} (\href{https://orcid.org/0000-0001-5296-5267}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Fabian Münch \email{fabian.felix.muench@uni-jena.de} (\href{https://orcid.org/0000-0001-5591-9901}{ORCID})
  \item Tobias Koch \email{tobias.koch@uni-jena.de} (\href{https://orcid.org/0000-0002-8143-3566}{ORCID})
  \item Jana Holtmann \email{jana.holtmann@uni-leipzig.de} (\href{https://orcid.org/0000-0002-7949-0772}{ORCID})
}

}
\keyword{internal}
