% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_fit.R
\name{mlts_fit}
\alias{mlts_fit}
\title{Fit Bayesian Multilevel Manifest or Latent Time-Series Models}
\usage{
mlts_fit(
  model,
  data = NULL,
  id,
  ts,
  covariates = NULL,
  outcomes = NULL,
  outcome_pred_btw = NULL,
  center_covs = T,
  time = NULL,
  tinterval,
  beep = NULL,
  days = NULL,
  n_overnight_NAs,
  na.rm = F,
  iter = 500,
  chains = 2,
  cores = 2,
  monitor_person_pars = F,
  fit_model = T,
  print_message = T,
  print_warning = T,
  ...
)
}
\arguments{
\item{model}{data.frame. Output of \code{\link[mlts]{mlts_model}} and related functions.}

\item{data}{An object of class data.frame (or one that can be coerced to that
class) containing data of all variables used in the model. Alternatively,
a list object with simulated data created by \code{mlts_sim} can be entered directly
and allows for comparison of estimates and true population paramter values used
in the data generation.}

\item{id}{Character. The variable in \code{data} that identifies the person or observational
unit. Not necessary when \code{data} is a list object of simulated data generated
with \code{mlts_sim}.}

\item{ts}{Character. The variable(s) in \code{data} that contain the time-series
construct(s). If multiple constructs are provided in the \code{model}, multiple entries
are necessary. Note that the order of variable names provided in \code{ts} has to match
the specification made in the \code{model}. E.g., if multiple constructs (e.g.,
\code{mlts_model(q = 2)}) are provided the order of variables names provided in \code{ts}
determines which construct is referred to as mu_1, phi(1)_11, etc..}

\item{covariates}{Named character vector. An optional named vector of
characters to refer to predictors of random effects as specified in the \code{model}.
Note that specifying \code{covariates} is only necessary if the respective
variable name(s) in \code{data} differ from the variables names specified in \code{model}.}

\item{outcomes}{Named character vector. Similar to \code{covariates}, an optional named vector of
characters to refer to outcome predicted by random effects as specified in the \code{model}.
Note that specifying \code{outcomes} is only necessary if the respective
variable name(s) in \code{data} differ from the outcome variable name(s) specified in \code{model}.}

\item{outcome_pred_btw}{Character.}

\item{center_covs}{Logical. Between-level covariates used as predictors of random effects
will be grand-mean centered before model fitting by default to allow interpretation
of fixed effects intercepts as mean population effects for average covariate values.
Setting \code{center_covs} to \code{FALSE} might be necessary when including categorical
predictors into the set of \code{covariates}. Note that in this case, additional
continuous covariates should be grand-mean centered prior to using \code{mlts_fit}.}

\item{time}{Character. The variable in \code{data} that contains the (continuous) time.}

\item{tinterval}{The step interval for approximation for a continuous time
dynamic model. The smaller the step interval, the better the approximation.}

\item{beep}{tba.}

\item{days}{tba.}

\item{n_overnight_NAs}{tba.}

\item{na.rm}{tba.}

\item{iter}{A positive integer specifying the number of iterations for each
chain (including 50\% used as warmup). The default is 500.}

\item{chains}{A positive integer specifying the number of Markov chains.
The default is 2.}

\item{cores}{The number of cores to use when executing the Markov chains in parallel.
The default is 2 (see \code{\link[rstan]{stan}}).}

\item{monitor_person_pars}{Logical. Should person parameters (i.e., values of the
latent variables) be stored? Default is FALSE.}

\item{fit_model}{Logical. Set to FALSE to avoid fitting the model which may be
helpful to inspect prepared data used for model estimation (default = TRUE).}

\item{print_message}{Logical. Print messages based on defined inputs (default = TRUE).}

\item{print_warning}{Logical. Print warnings based on defined inputs (default = TRUE).}

\item{...}{Additional arguments passed to \code{\link[rstan]{sampling}}.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Fit Bayesian Multilevel Manifest or Latent Time-Series Models
}
\examples{
\donttest{
 # build simple mlts model for two time-series variables
 model <- mlts_model(q = 2)

 # fit model with (artificial) dataset ts_data
 fit <- mlts_fit(
   model = model,
   data = ts_data,
   ts = c("Y1", "Y2"), # time-series variables
   id = "ID", # identifier variable
   tinterval = 1 # interval for approximation of continuous-time dynamic model,
 )

 # inspect model summary
 summary(fit)
}

}
