% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_fit.R
\name{mlts_fit}
\alias{mlts_fit}
\title{Title}
\usage{
mlts_fit(
  VARmodel,
  data = NULL,
  id,
  ts.ind,
  covariates = NULL,
  outcomes = NULL,
  outcome.pred.btw = NULL,
  center.covs = T,
  std.outcome = T,
  time = NULL,
  tinterval,
  beep = NULL,
  days = NULL,
  n_overnight_NAs,
  na.rm = F,
  iter = 500,
  chains = 2,
  cores = 2,
  monitor.person.pars = F,
  fit.model = T,
  printMessage = T,
  printWarning = T,
  ...
)
}
\arguments{
\item{VARmodel}{data.frame. Output of VARmodel-Functions.}

\item{data}{An object of class data.frame (or one that can be coerced to that
class) containing data of all variables used in the model.}

\item{ts.ind}{data.frame. Data input.}

\item{covariates}{data.frame. Data input.}

\item{outcomes}{data.frame. Data input.}

\item{outcome.pred.btw}{data.frame. Data input.}

\item{center.covs}{data.frame. Data input.}

\item{std.outcome}{data.frame. Data input.}

\item{iter}{A positive integer specifying the number of iterations for each
chain (including 50\% used as warmup). The default is 500.}

\item{chains}{A positive integer specifying the number of Markov chains. The default is 2.}

\item{cores}{The number of cores to use when executing the Markov chains in parallel. The default is 2 (see \code{\link[rstan]{stan}}).}

\item{monitor.person.pars}{data.frame. Data input.}

\item{fit.model}{logical. Set to FALSE to avoid fitting the model which may be
helpful to inspect prepared data used for model estimation (default = T).}

\item{printMessage}{logical. Print messages based on defined inputs (default = TRUE).}

\item{printWarning}{logical. Print messages based on defined inputs (default = TRUE).}

\item{...}{Additional arguments passed to \code{\link[rstan]{sampling}}.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
