% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_fit.R
\name{mlts_fit}
\alias{mlts_fit}
\title{Fit Bayesian Multilevel Manifest or Latent Time-Series Models}
\usage{
mlts_fit(
  model,
  data = NULL,
  id,
  ts,
  covariates = NULL,
  outcomes = NULL,
  outcome_pred_btw = NULL,
  center_covs = T,
  time = NULL,
  tinterval,
  beep = NULL,
  days = NULL,
  n_overnight_NAs,
  na.rm = F,
  iter = 500,
  chains = 2,
  cores = 2,
  monitor_person_pars = F,
  fit_model = T,
  print_message = T,
  print_warning = T,
  ...
)
}
\arguments{
\item{model}{data.frame. Output of \code{\link[mlts]{mlts_model}} and related functions.}

\item{data}{An object of class data.frame (or one that can be coerced to that
class) containing data of all variables used in the model.}

\item{id}{Character. The variable in \code{data} that identifies the person or observational
unit.}

\item{ts}{Character. The variable(s) in \code{data} that
contain the time-series construct. If multiple variables are provided in a
character vector, a vector autoregressive model is fit.}

\item{covariates}{Character. The covariate(s) in \code{data} used for prediction of
random effects.}

\item{outcomes}{Character. The outcome(s) in \code{data} that should be
predicted by random effects.}

\item{outcome_pred_btw}{Character.}

\item{center_covs}{Logical. Should covariates be centered before inclusion
in the model? Defaults to \code{TRUE}.}

\item{time}{Character. The variable in \code{data} that contains the (continuous) time.}

\item{tinterval}{The step interval for approximation for a continuous time
dynamic model. The smaller the step interval, the better the approximation.}

\item{beep}{tba.}

\item{days}{tba.}

\item{n_overnight_NAs}{tba.}

\item{na.rm}{tba.}

\item{iter}{A positive integer specifying the number of iterations for each
chain (including 50\% used as warmup). The default is 500.}

\item{chains}{A positive integer specifying the number of Markov chains.
The default is 2.}

\item{cores}{The number of cores to use when executing the Markov chains in parallel.
The default is 2 (see \code{\link[rstan]{stan}}).}

\item{monitor_person_pars}{Logical. Should person parameters (i.e., values of the
latent variables) be stored? Default is FALSE.}

\item{fit_model}{Logical. Set to FALSE to avoid fitting the model which may be
helpful to inspect prepared data used for model estimation (default = TRUE).}

\item{print_message}{Logical. Print messages based on defined inputs (default = TRUE).}

\item{print_warning}{Logical. Print warnings based on defined inputs (default = TRUE).}

\item{...}{Additional arguments passed to \code{\link[rstan]{sampling}}.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Fit Bayesian Multilevel Manifest or Latent Time-Series Models
}
\examples{
\donttest{
 # build simple mlts model for two time-series variables
 model <- mlts_model(q = 2)

 # fit model with (artificial) dataset ts_data
 fit <- mlts_fit(
   model = model,
   data = ts_data,
   ts = c("Y1", "Y2"), # time-series variables
   id = "ID", # identifier variable
   tinterval = 1 # interval for approximation of continuous-time dynamic model,
 )

 # inspect model summary
 summary(fit)
}

}
