% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_model_betw.R
\name{mlts_model_betw}
\alias{mlts_model_betw}
\title{Title}
\usage{
mlts_model_betw(
  model,
  ranef_pred = NULL,
  out_pred = NULL,
  out_pred_add_btw = NULL
)
}
\arguments{
\item{model}{data.frame. Output of \code{model()}-functions.}

\item{ranef_pred}{character vector or a named list. Include between-level covariate(s)
as predictor(s) of all random effects in \code{model} by entering a vector of unique variable
names. Alternatively, to include between-level covariates or differing sets of
between-level covariates as predictors of specific random effects, a named
list (using the \code{param}-labels in \code{model}) can be entered (see details).
Note that if a named list is provided, all names that do not match random
parameters in \code{model} will be ignored.}

\item{out_pred}{character vector or a named list. Include between-level covariate(s)
as predictor(s) of all random effects in \code{model} by entering a vector of unique variable
names. Alternatively, to include between-level covariates or differing sets of
between-level covariates as predictors of specific random effects, a named
list (using the \code{param}-labels in \code{model}) can be entered (see details).}

\item{out_pred_add_btw}{tba}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
