% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_model_check.R
\name{mlts_model_check}
\alias{mlts_model_check}
\title{Title}
\usage{
mlts_model_check(model, data, ts.ind, covariates, outcomes)
}
\arguments{
\item{model}{data.frame. Output of model-Functions.}

\item{data}{data.frame. Data input.}

\item{ts.ind}{tba}

\item{covariates}{tba}

\item{outcomes}{tba.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
