% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_model_constraint.R
\name{mlts_model_constraint}
\alias{mlts_model_constraint}
\title{Title}
\usage{
mlts_model_constraint(
  model,
  fix_dynamics = F,
  fix_inno_vars = F,
  fix_inno_covs = F,
  inno_covs_zero = F,
  fixef_zero = NULL,
  ranef_zero = NULL
)
}
\arguments{
\item{model}{data.frame. Output of model-Functions.}

\item{fix_dynamics}{logical. Fix all random effect variances (except those
of individual traits) to zero.}

\item{fix_inno_vars}{logical. Set all innovation variances to a constant value.}

\item{fix_inno_covs}{logical. Set all innovation covariances to a constant value.}

\item{inno_covs_zero}{logical. Set to TRUE to treat all innovations as independent.}

\item{fixef_zero}{character. A character vector to indemodel which fimodeled model parameters
should be fimodeled to zero (Note: this results in removing the random effect
variance of the respective parameter).}

\item{ranef_zero}{logical. Set to TRUE to treat all innovations as independent.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
