% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_model_eval.R
\name{mlts_model_eval}
\alias{mlts_model_eval}
\title{Title}
\usage{
mlts_model_eval(model)
}
\arguments{
\item{model}{data.frame. Output of model-Functions.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
