% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_plot.R
\name{mlts_plot}
\alias{mlts_plot}
\title{Plot results of mlts}
\usage{
mlts_plot(
  fit,
  type = c("fe", "re", "re.cor"),
  sort.est = NULL,
  xlab = NULL,
  ylab = NULL,
  n.facet.col = 1,
  pt.size = 1,
  pt.col = "black",
  pt.shape = 1,
  err.bar.col = "black",
  err.bar.width = 0.3,
  add.true = FALSE,
  true.col = "red",
  true.shape = 22,
  true.size = 1,
  hide.xaxis.text = T
)
}
\arguments{
\item{fit}{An object of class \code{mlts.fit}}

\item{what}{one of "fixef", "ranef", "raneftab"}
}
\value{
A \code{list} that can be passed to \code{stan()}.
}
\description{
Plot results of mlts
}
\examples{
TBA
}
