% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_plot.R
\name{mlts_plot}
\alias{mlts_plot}
\title{Plot results of mlts}
\usage{
mlts_plot(
  fit,
  type = c("fe", "re", "re.cor"),
  bpe = c("median", "mean"),
  sort.est = NULL,
  xlab = NULL,
  ylab = NULL,
  facet.ncol = 1,
  dot.size = 1,
  dot.color = "black",
  dot.shape = 1,
  err.bar.col = "black",
  err.bar.width = 0.3,
  add.true = FALSE,
  true.col = "red",
  true.shape = 22,
  true.size = 1,
  hide.xaxis.text = T,
  par_labels = NULL,
  labels.as.expression = F
)
}
\arguments{
\item{fit}{An object of class \code{mlts.fit}}

\item{type}{Type of plot.
type = "fixef" (Default)
Forest-plot of model coefficients.
type = "ranef"
Plot of individual (random) effects
type = "raneftab"
(not yet implemented)}

\item{bpe}{The Bayesian point estimate is, by default, the median of the
posterior distribution (\code{bpe = "median"}). Set \code{bpe = "mean"} to use
the mean of the posterior distribution as point estimates.}

\item{sort.est}{Add parameter label for sorting of random effects.}

\item{xlab}{Title for the x axis.}

\item{ylab}{Title for the y axis.}

\item{facet.ncol}{Number of facet columns (see \code{ggplot2::facet_grid}).}

\item{dot.size}{Numeric, size of the dots that indicate the point estimates.}

\item{dot.color}{Character vector, indicating the color of the point estimates.}

\item{dot.shape}{Numeric, shape of the dots that indicate the point estimates.}

\item{err.bar.col}{integer.}

\item{err.bar.width}{integer.}

\item{add.true}{integer.}

\item{true.col}{integer.}

\item{true.shape}{integer.}

\item{true.size}{integer.}

\item{hide.xaxis.text}{logical.}

\item{par_labels}{integer.}

\item{labels.as.expression}{logical.}
}
\value{
Returns a \code{ggplot}-object .
}
\description{
Plot results of mlts
}
\examples{
TBA
}
