% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlts_sim.R
\name{mlts_sim}
\alias{mlts_sim}
\title{Simulate data from mlts model}
\usage{
mlts_sim(
  model,
  default = F,
  N,
  TP,
  burn.in = 50,
  seed = NULL,
  btw.var.sds = NULL
)
}
\arguments{
\item{model}{data.frame. Output of \code{\link[mlts]{mlts_model}}.}

\item{default}{logical. If set to \code{TRUE}, default prior specifications are
added.}

\item{N}{integer. Number of observational units.}

\item{TP}{integer. Number of measurements per observational unit.}

\item{burn.in}{integer.}

\item{seed}{integer.}

\item{btw.var.sds}{integer.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Simulate data from mlts model
}
