% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for Stan}
\usage{
prepare_data(
  data,
  id,
  ts_ind,
  time,
  tinterval,
  beep,
  days,
  n_overnight_NAs,
  na.rm = FALSE,
  outcomes = NULL,
  covariates = NULL,
  outcome.pred.btw = NULL
)
}
\arguments{
\item{data}{An object of class \code{data.frame} (or one that can be coerced
to that class) in long format.}

\item{id}{character. The variable in \code{data} that identifies the person or
observational unit.}

\item{ts_ind}{character. The variable(s) in \code{data} that contain(s) the time series data.}

\item{time}{character. The variable in \code{data} that contains the (continuous) time (as
string).}

\item{tinterval}{The step interval for approximation for a continuous time DSEM.
The smaller the step interval, the better the approximation.}

\item{beep}{character. The variable in \code{data} that contains the running
beep from 1 to TP for each person.}

\item{days}{Optional. If a running beep identifier is provided via the \code{beep} argument and
observations are nested within days (or similar grouping unit), the variable
in \code{data} that contains the day identifier can be added to correct for overnight lags (see Details).}

\item{na.rm}{logical. As default option missing values remain in the data and
will be imputed during model estimation. Set to \code{TRUE} to remove all rows with
missing values in variables given in \code{ts_ind}.}

\item{n_overnight.NAs}{Optional. The number of \code{NA} rows to add after the last observation of each day (if \code{day} is provided).}
}
\value{
A \code{list} that can be passed to \code{stan()}.
}
\description{
Prepare data for Stan
}
\examples{
TBA
}
