% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mltsfit}
\alias{summary.mltsfit}
\title{Create a summary of a fitted model with class \code{mltsfit}}
\usage{
\method{summary}{mltsfit}(object, priors = FALSE, se = FALSE, prob = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{mltsfit}.}

\item{priors}{Logical. Should priors be included in the summary?
Defaults to \code{FALSE}.}

\item{se}{Logical. Should the Monte Carlo Standard Error be included
in the summary? Defaults to \code{FALSE}.}

\item{prob}{A value between 0 and 1 to indicate the width of the credible
interval. Default is .95.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Create a summary of a fitted model with class \code{mltsfit}
}
