% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARmodelBuild.R
\name{VARmodelBuild}
\alias{VARmodelBuild}
\title{Title}
\usage{
VARmodelBuild(
  q,
  p = NULL,
  maxLag = c(1, 2, 3),
  btw.factor = TRUE,
  btw.model = NULL,
  FixDynamics = F,
  FixInnoVars = F,
  FixInnoCovs = F,
  InnoCovsZero = F,
  FEis0 = NULL,
  REis0 = NULL,
  RE.pred = NULL,
  out.pred = NULL,
  out.pred.add.btw = NULL
)
}
\arguments{
\item{q}{integer. The number of time-varying constructs.}

\item{p}{integer. For multiple-indicator models, specify a vector of length
\code{q} with the number of manifest indicators per construct.}

\item{btw.factor}{Logical. If \code{TRUE} (the default), a common between-level factor
is modeled across all indicator variables. If \code{FALSE}, instead of a between-level
factor, indicator mean levels will be included as individual (random) effects stemming
from a joint multivariate normal distribution.}

\item{btw.model}{A list to indicate for which manifest indicator variables a common
between-level factor should be modeled (see Details for detailed instructions).
At this point restricted to one factor per latent construct.}

\item{FEis0}{character. A character vector to indeVARmodel which fiVARmodeled model parameters
should be fiVARmodeled to zero (Note: this results in removing the random effect
variance of the respective parameter).}

\item{REis0}{logical. Set to TRUE to treat all innovations as independent.}

\item{RE.pred}{character vector or a named list. Include between-level covariate(s)
as predictor(s) of all random effects in \code{VARmodel} by entering a vector of unique variable
names. Alternatively, to include between-level covariates or differing sets of
between-level covariates as predictors of specific random effects, a named
list (using the \code{param}-labels in \code{VARmodel}) can be entered (see details).
Note that if a named list is provided, all names that do not match random
parameters in \code{VARmodel} will be ignored.}

\item{out.pred}{character vector or a named list. Include between-level covariate(s)
as predictor(s) of all random effects in \code{VARmodel} by entering a vector of unique variable
names. Alternatively, to include between-level covariates or differing sets of
between-level covariates as predictors of specific random effects, a named
list (using the \code{param}-labels in \code{VARmodel}) can be entered (see details).}

\item{FiVARmodeledInnoVars}{logical. FiVARmodel all random effect variances (eVARmodelcept those
of individual traits) to zero.}

\item{FiVARmodeledCovs}{logical. Set all innovation covariances to a constant value.}

\item{CovsZero}{logical. Set to TRUE to treat all innovations as independent.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
