% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARprepare.R
\name{VARprepare}
\alias{VARprepare}
\title{Title}
\usage{
VARprepare(
  VARmodel,
  data,
  ts.ind,
  covariates = NULL,
  outcomes = NULL,
  outcome.pred.btw = NULL,
  center.covs = T,
  std.outcome = T
)
}
\arguments{
\item{VARmodel}{data.frame. Output of VARmodel-Functions.}

\item{data}{data.frame. Data input.}

\item{ts.ind}{tba.}

\item{covariates}{tba.}

\item{outcomes}{tba.}

\item{outcome.pred.btw}{tba.}

\item{center.covs}{tba.}

\item{std.outcome}{tba.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
