% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARprepare.R
\name{VARprepare}
\alias{VARprepare}
\title{Title}
\usage{
VARprepare(
  model,
  data,
  ts,
  covariates = NULL,
  outcomes = NULL,
  outcome_pred_btw = NULL,
  center_covs = T,
  std_outcome = T
)
}
\arguments{
\item{model}{data.frame. Output of model-Functions.}

\item{data}{data.frame. Data input.}

\item{ts}{tba.}

\item{covariates}{tba.}

\item{outcomes}{tba.}

\item{outcome_pred_btw}{tba.}

\item{center_covs}{tba.}

\item{std_outcome}{tba.}
}
\value{
An object of class \code{data.frame}.
}
\description{
Title
}
